Update and rename logistic regression model on social media ads to model to predict effect of social media ads on consumers (buy or not buy)
This is a model predicting whether users would purchase a car advertised to them on social media based on their age and salary. test size is 0.25. and logistic regression model is used because it gave best results after testing on the confusion matrix (results of confusion matrix (true positives =24, true negatives=65, FP =3 & FN=8)